feat: Add AWS resources for web server deployment
A new Terraform configuration file, `main.tf`, was added in the `deployments` directory. This file defines an AWS security group and an AWS instance for a web server. The security group (`web_server_sg`) allows for HTTP and SSH traffic. The AWS instance employs a t2.micro instance running a dockerized Nginx web server. This setup prepares the AWS environment for the deployment of a web server.